refactor some of ht eutility code
