+++ ferritin-featurizers/src/models/ligandmpnn/utils.rs
@@ -46,202 +46,6 @@ pub fn cat_neighbors_nodes(
     Tensor::cat(&[h_neighbors, &h_nodes_gathered], D::Minus1)
 }
 
-fn restype_1to3(code: char) -> &'static str {
-    match code {
-        'A' => "ALA",
-        'R' => "ARG",
-        'N' => "ASN",
-        'D' => "ASP",
-        'C' => "CYS",
-        'Q' => "GLN",
-        'E' => "GLU",
-        'G' => "GLY",
-        'H' => "HIS",
-        'I' => "ILE",
-        'L' => "LEU",
-        'K' => "LYS",
-        'M' => "MET",
-        'F' => "PHE",
-        'P' => "PRO",
-        'S' => "SER",
-        'T' => "THR",
-        'W' => "TRP",
-        'Y' => "TYR",
-        'V' => "VAL",
-        _ => "UNK",
-    }
-}
-
-fn restype_3to1(restype: &str) -> char {
-    match restype {
-        "ALA" => 'A',
-        "ARG" => 'R',
-        "ASN" => 'N',
-        "ASP" => 'D',
-        "CYS" => 'C',
-        "GLN" => 'Q',
-        "GLU" => 'E',
-        "GLY" => 'G',
-        "HIS" => 'H',
-        "ILE" => 'I',
-        "LEU" => 'L',
-        "LYS" => 'K',
-        "MET" => 'M',
-        "PHE" => 'F',
-        "PRO" => 'P',
-        "SER" => 'S',
-        "THR" => 'T',
-        "TRP" => 'W',
-        "TYR" => 'Y',
-        "VAL" => 'V',
-        _ => 'X',
-    }
-}
-
-pub fn restype_str_to_int(code: char) -> i32 {
-    match code.to_ascii_uppercase() {
-        'A' => 0,
-        'C' => 1,
-        'D' => 2,
-        'E' => 3,
-        'F' => 4,
-        'G' => 5,
-        'H' => 6,
-        'I' => 7,
-        'K' => 8,
-        'L' => 9,
-        'M' => 10,
-        'N' => 11,
-        'P' => 12,
-        'Q' => 13,
-        'R' => 14,
-        'S' => 15,
-        'T' => 16,
-        'V' => 17,
-        'W' => 18,
-        'Y' => 19,
-        'X' => 20,
-        _ => 20, // Default to 'X' (20) for any unrecognized character
-    }
-}
-
-fn restype_int_to_str(code: i32) -> &'static str {
-    match code {
-        0 => "A",
-        1 => "C",
-        2 => "D",
-        3 => "E",
-        4 => "F",
-        5 => "G",
-        6 => "H",
-        7 => "I",
-        8 => "K",
-        9 => "L",
-        10 => "M",
-        11 => "N",
-        12 => "P",
-        13 => "Q",
-        14 => "R",
-        15 => "S",
-        16 => "T",
-        17 => "V",
-        18 => "W",
-        19 => "Y",
-        _ => "X", // Default to 'X' for any unrecognized integer
-    }
-}
-
-#[rustfmt::skip]
-fn restype_name_to_atom14_names(restype: &str) -> [&'static str; 14] {
-    match restype {
-        "ALA" => ["N", "CA", "C", "O", "CB", "", "", "", "", "", "", "", "", "",],
-        "ARG" => ["N", "CA", "C", "O", "CB", "CG", "CD", "NE", "CZ", "NH1", "NH2", "", "", "",],
-        "ASN" => ["N", "CA", "C", "O", "CB", "CG", "OD1", "ND2", "", "", "", "", "", "",],
-        "ASP" => ["N", "CA", "C", "O", "CB", "CG", "OD1", "OD2", "", "", "", "", "", "",],
-        "CYS" => ["N", "CA", "C", "O", "CB", "SG", "", "", "", "", "", "", "", "",],
-        "GLN" => ["N", "CA", "C", "O", "CB", "CG", "CD", "OE1", "NE2", "", "", "", "", "",],
-        "GLU" => ["N", "CA", "C", "O", "CB", "CG", "CD", "OE1", "OE2", "", "", "", "", "",],
-        "GLY" => ["N", "CA", "C", "O", "", "", "", "", "", "", "", "", "", "",],
-        "HIS" => ["N", "CA", "C", "O", "CB", "CG", "ND1", "CD2", "CE1", "NE2", "", "", "", "",],
-        "ILE" => ["N", "CA", "C", "O", "CB", "CG1", "CG2", "CD1", "", "", "", "", "", "",],
-        "LEU" => ["N", "CA", "C", "O", "CB", "CG", "CD1", "CD2", "", "", "", "", "", "",],
-        "LYS" => ["N", "CA", "C", "O", "CB", "CG", "CD", "CE", "NZ", "", "", "", "", "",],
-        "MET" => ["N", "CA", "C", "O", "CB", "CG", "SD", "CE", "", "", "", "", "", "",],
-        "PHE" => ["N", "CA", "C", "O", "CB", "CG", "CD1", "CD2", "CE1", "CE2", "CZ", "", "", "",],
-        "PRO" => ["N", "CA", "C", "O", "CB", "CG", "CD", "", "", "", "", "", "", "",],
-        "SER" => ["N", "CA", "C", "O", "CB", "OG", "", "", "", "", "", "", "", "",],
-        "THR" => ["N", "CA", "C", "O", "CB", "OG1", "CG2", "", "", "", "", "", "", "",],
-        "TRP" => ["N", "CA", "C", "O", "CB", "CG", "CD1", "CD2", "CE2", "CE3", "NE1", "CZ2", "CZ3", "CH2",],
-        "TYR" => ["N", "CA", "C", "O", "CB", "CG", "CD1", "CD2", "CE1", "CE2", "CZ", "OH", "", "",],
-        "VAL" => ["N", "CA", "C", "O", "CB", "CG1", "CG2", "", "", "", "", "", "", "",],
-        _ => ["", "", "", "", "", "", "", "", "", "", "", "", "", "",], // UNK or any other case
-    }
-}
-
-#[rustfmt::skip]
-fn atom_order(atom: &str) -> i32 {
-    match atom {
-        "N" => 0,
-        "CA" => 1,
-        "C" => 2,
-        "CB" => 3,
-        "O" => 4,
-        "CG" => 5,
-        "CG1" => 6,
-        "CG2" => 7,
-        "OG" => 8,
-        "OG1" => 9,
-        "SG" => 10,
-        "CD" => 11,
-        "CD1" => 12,
-        "CD2" => 13,
-        "ND1" => 14,
-        "ND2" => 15,
-        "OD1" => 16,
-        "OD2" => 17,
-        "SD" => 18,
-        "CE" => 19,
-        "CE1" => 20,
-        "CE2" => 21,
-        "CE3" => 22,
-        "NE" => 23,
-        "NE1" => 24,
-        "NE2" => 25,
-        "OE1" => 26,
-        "OE2" => 27,
-        "CH2" => 28,
-        "NH1" => 29,
-        "NH2" => 30,
-        "OH" => 31,
-        "CZ" => 32,
-        "CZ2" => 33,
-        "CZ3" => 34,
-        "NZ" => 35,
-        "OXT" => 36,
-        _ => -1, // Return -1 for unknown atoms
-    }
-}
-
-const ALPHABET: [char; 21] = [
-    'A', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'K', 'L', 'M', 'N', 'P', 'Q', 'R', 'S', 'T', 'V', 'W',
-    'Y', 'X',
-];
-
-const ELEMENT_LIST: [&str; 118] = [
-    "H", "He", "Li", "Be", "B", "C", "N", "O", "F", "Ne", "Na", "Mg", "Al", "Si", "P", "S", "Cl",
-    "Ar", "K", "Ca", "Sc", "Ti", "V", "Cr", "Mn", "Fe", "Co", "Ni", "Cu", "Zn", "Ga", "Ge", "As",
-    "Se", "Br", "Kr", "Rb", "Sr", "Y", "Zr", "Nb", "Mo", "Tc", "Ru", "Rh", "Pd", "Ag", "Cd", "In",
-    "Sn", "Sb", "Te", "I", "Xe", "Cs", "Ba", "La", "Ce", "Pr", "Nd", "Pm", "Sm", "Eu", "Gd", "Tb",
-    "Dy", "Ho", "Er", "Tm", "Yb", "Lu", "Hf", "Ta", "W", "Re", "Os", "Ir", "Pt", "Au", "Hg", "Tl",
-    "Pb", "Bi", "Po", "At", "Rn", "Fr", "Ra", "Ac", "Th", "Pa", "U", "Np", "Pu", "Am", "Cm", "Bk",
-    "Cf", "Es", "Fm", "Md", "No", "Lr", "Rf", "Db", "Sg", "Bh", "Hs", "Mt", "Ds", "Rg", "Cn", "Nh",
-    "Fl", "Mc", "Lv", "Ts", "Og",
-];
-
-// element_list = [item.upper() for item in element_list]
-// element_dict = dict(zip(element_list, range(1,len(element_list))))
-// element_dict_rev = dict(zip(range(1, len(element_list)), element_list))
-
 fn get_seq_rec(s: &Tensor, s_pred: &Tensor, mask: &Tensor) -> Result<Tensor> {
     // S: true sequence shape=[batch, length]
     // S_pred: predicted sequence shape=[batch, length]
@@ -293,138 +97,6 @@ fn get_score(s: &Tensor, log_probs: &Tensor, mask: &Tensor) -> Result<(Tensor, T
 
     Ok((average_loss, loss_per_residue))
 }
-
-// Todo: finish this. will require somethign like prody....
-// pub fn write_full_pdb(
-//     save_path: &str,
-//     x: &Tensor,
-//     x_m: &Tensor,
-//     b_factors: &Tensor,
-//     r_idx: &Tensor,
-//     chain_letters: &Tensor,
-//     s: &Tensor,
-//     other_atoms: Option<&Tensor>,
-//     icodes: Option<&Tensor>,
-//     force_hetatm: bool,
-// ) -> Result<()> {
-//     //     save_path : path where the PDB will be written to
-//     //     X : protein atom xyz coordinates shape=[length, 14, 3]
-//     //     X_m : protein atom mask shape=[length, 14]
-//     //     b_factors: shape=[length, 14]
-//     //     R_idx: protein residue indices shape=[length]
-//     //     chain_letters: protein chain letters shape=[length]
-//     //     S : protein amino acid sequence shape=[length]
-//     //     other_atoms: other atoms parsed by prody
-//     //     icodes: a list of insertion codes for the PDB; e.g. antibody loops
-//     //     """
-
-//     let s_str: Vec<&str> = s
-//         .iter()
-//         .map(|&aa| restype_int_to_str(aa))
-//         .map(restype_1to3)
-//         .collect();
-//     let mut x_list = Vec::new();
-//     let mut b_factor_list = Vec::new();
-//     let mut atom_name_list = Vec::new();
-//     let mut element_name_list = Vec::new();
-//     let mut residue_name_list = Vec::new();
-//     let mut residue_number_list = Vec::new();
-//     let mut chain_id_list = Vec::new();
-//     let mut icodes_list = Vec::new();
-
-//     for (i, aa) in s_str.iter().enumerate() {
-//         let sel = x_m.get(i)?.to_dtype(DType::I32)?.eq(&1)?;
-//         let total = sel.sum_all()?.to_scalar::<i32>()?;
-//         let tmp = Tensor::from_slice(&restype_name_to_atom14_names(aa))?.masked_select(&sel)?;
-//         x_list.push(x.get(i)?.masked_select(&sel)?);
-//         b_factor_list.push(b_factors.get(i)?.masked_select(&sel)?);
-//         atom_name_list.push(tmp.clone());
-//         element_name_list.extend(tmp.iter().map(|&atom| &atom[..1]));
-//         residue_name_list.extend(std::iter::repeat(aa).take(total as usize));
-//         residue_number_list.extend(std::iter::repeat(r_idx.get(i)?).take(total as usize));
-//         chain_id_list.extend(std::iter::repeat(chain_letters.get(i)?).take(total as usize));
-//         icodes_list.extend(std::iter::repeat(icodes.get(i)?).take(total as usize));
-//     }
-
-//     let x_stack = Tensor::cat(&x_list, 0)?;
-//     let b_factor_stack = Tensor::cat(&b_factor_list, 0)?;
-//     let atom_name_stack = Tensor::cat(&atom_name_list, 0)?;
-
-//     let mut protein = prody::AtomGroup::new();
-//     protein.set_coords(&x_stack)?;
-//     protein.set_betas(&b_factor_stack)?;
-//     protein.set_names(&atom_name_stack)?;
-//     protein.set_resnames(&residue_name_list)?;
-//     protein.set_elements(&element_name_list)?;
-//     protein.set_occupancies(&Tensor::ones(x_stack.shape()[0])?)?;
-//     protein.set_resnums(&residue_number_list)?;
-//     protein.set_chids(&chain_id_list)?;
-//     protein.set_icodes(&icodes_list)?;
-
-//     if let Some(other_atoms) = other_atoms {
-//         let mut other_atoms_g = prody::AtomGroup::new();
-//         other_atoms_g.set_coords(&other_atoms.get_coords()?)?;
-//         other_atoms_g.set_names(&other_atoms.get_names()?)?;
-//         other_atoms_g.set_resnames(&other_atoms.get_resnames()?)?;
-//         other_atoms_g.set_elements(&other_atoms.get_elements()?)?;
-//         other_atoms_g.set_occupancies(&other_atoms.get_occupancies()?)?;
-//         other_atoms_g.set_resnums(&other_atoms.get_resnums()?)?;
-//         other_atoms_g.set_chids(&other_atoms.get_chids()?)?;
-//         if force_hetatm {
-//             other_atoms_g.set_flags("hetatm", &other_atoms.get_flags("hetatm")?)?;
-//         }
-//         prody::write_pdb(save_path, &(protein + other_atoms_g))?;
-//     } else {
-//         prody::write_pdb(save_path, &protein)?;
-//     }
-// }
-
-// Todo: finish this
-// pub fn get_aligned_coordinates(
-//     protein_atoms: AtomCollection,
-//     ca_dict: &HashMap<String, usize>,
-//     atom_name: &str,
-// ) -> Result<(Tensor, Tensor)> {
-//     let atom_atoms = protein_atoms.select(&format!("name {}", atom_name));
-
-//     let (atom_coords, atom_resnums, atom_chain_ids, atom_icodes) = if let Some(atoms) = atom_atoms {
-//         (
-//             atoms.get_coords()?,
-//             atoms.get_resnums()?,
-//             atoms.get_chids()?,
-//             atoms.get_icodes()?,
-//         )
-//     } else {
-//         (
-//             Tensor::zeros((0, 3), DType::F32, &Device::Cpu)?,
-//             Tensor::zeros((0,), DType::I64, &Device::Cpu)?,
-//             Tensor::zeros((0,), DType::I64, &Device::Cpu)?,
-//             Tensor::zeros((0,), DType::I64, &Device::Cpu)?,
-//         )
-//     };
-
-//     let mut atom_coords_ = Tensor::zeros((ca_dict.len(), 3), DType::F32, &Device::Cpu)?;
-//     let mut atom_coords_m = Tensor::zeros((ca_dict.len(),), DType::I32, &Device::Cpu)?;
-
-//     if let Some(atoms) = atom_atoms {
-//         for i in 0..atom_resnums.dim(0)? {
-//             let code = format!(
-//                 "{}_{}{}",
-//                 atom_chain_ids.get(i)?,
-//                 atom_resnums.get(i)?,
-//                 atom_icodes.get(i)?
-//             );
-//             if let Some(&idx) = ca_dict.get(&code) {
-//                 atom_coords_
-//                     .slice_mut(1, idx, idx + 1)?
-//                     .copy_(&atom_coords.slice(1, i, i + 1)?)?;
-//                 atom_coords_m.set(idx, 1)?;
-//             }
-//         }
-//     }
-
-//     Ok((atom_coords_, atom_coords_m))
-// }
 
 // pub fn get_nearest_neighbours(
 //     cb: &Tensor,
